updating tests

+++ src/test/java/org/abstractj/kalium/publickey/AuthenticatedEncryptionTest.java
@@ -13,20 +13,20 @@ public class AuthenticatedEncryptionTest {
     @Test
     public void testAuthenticatedEncryption() {
         Sodium sodium= NaCl.sodium();
-        int alice_publickeylen=Sodium.crypto_box_publickeybytes();
-        int alice_privatekeylen=Sodium.crypto_box_secretkeybytes();
-        byte[] alice_public_key=new byte[alice_publickeylen];
-        byte[] alice_private_key=new byte[alice_privatekeylen];
+        long alice_publickeylen=Sodium.crypto_box_publickeybytes();
+        long alice_privatekeylen=Sodium.crypto_box_secretkeybytes();
+        byte[] alice_public_key=new byte[(int)alice_publickeylen];
+        byte[] alice_private_key=new byte[(int)alice_privatekeylen];
         System.out.println("Generating keypair");
         int ret=Sodium.crypto_box_keypair(alice_public_key,alice_private_key);
         Assert.assertEquals(0,ret);
         System.out.println(ret);
         System.out.println("Generated keyapir");
 
-        int bob_publickeylen=Sodium.crypto_box_publickeybytes();
-        int bob_privatekeylen=Sodium.crypto_box_secretkeybytes();
-        byte[] bob_public_key=new byte[bob_publickeylen];
-        byte[] bob_private_key=new byte[bob_privatekeylen];
+        long bob_publickeylen=Sodium.crypto_box_publickeybytes();
+        long bob_privatekeylen=Sodium.crypto_box_secretkeybytes();
+        byte[] bob_public_key=new byte[(int)bob_publickeylen];
+        byte[] bob_private_key=new byte[(int)bob_privatekeylen];
         System.out.println("Generating keypair");
         ret=Sodium.crypto_box_keypair(bob_public_key,bob_private_key);
         Assert.assertEquals(0,ret);
@@ -35,17 +35,17 @@ public class AuthenticatedEncryptionTest {
 
 
         byte[] message="test".getBytes();
-        int noncelen=Sodium.crypto_box_noncebytes();
-        byte[] nonce=new byte[noncelen];
-        int ciphertextlen=Sodium.crypto_box_macbytes()+message.length;
-        byte[] ciphertext=new byte[ciphertextlen];
+        long noncelen=Sodium.crypto_box_noncebytes();
+        byte[] nonce=new byte[(int)noncelen];
+        long ciphertextlen=Sodium.crypto_box_macbytes()+message.length;
+        byte[] ciphertext=new byte[(int)ciphertextlen];
 
-        Sodium.randombytes_buf(nonce,noncelen);
+        Sodium.randombytes_buf(nonce,(int)noncelen);
         ret=Sodium.crypto_box_easy(ciphertext,message,message.length,nonce,bob_public_key,alice_private_key);
         Assert.assertEquals(0,ret);
 
         byte[] decrypted=new byte[message.length];
-        ret=Sodium.crypto_box_open_easy(decrypted, ciphertext, ciphertextlen, nonce,
+        ret=Sodium.crypto_box_open_easy(decrypted, ciphertext, (int)ciphertextlen, nonce,
                 alice_public_key, bob_private_key);
         Assert.assertEquals(0,ret);
         System.out.println("Recovered message="+new String(decrypted));
